Refactor System called by compiled code to $traceurRuntime functions
Avoid calls to $traceurRuntime from within compiler

Rename ModuleStore -> ModuleStoreImpl and forward module ModuleStore to it
